Updated schedule to match actual topics covered
Instead of a full day of BDD with RSpec, we did RSpec with just Ruby in the morning, and then RSpec Rails in the afternoon (covering just models). I demonstrated how to test your models and then gave everyone time to try to replicate that process. The dusk lesson on Wednesday afternoon is now linked to the repo with the example I created in class (rspec_rails_example)