Fixed week 11 schedule structure.
Not sure what was wrong with it, but it seemed to include some topics from week 10 and I think just had some formatting errors.

Also added in a description about the week, the rules, and project 3 info. Reformatted schedule to reflect the fact that we're breaking the class into two groups so that it will hopefully be easier to read.